feat: optimize inject logic
