+++ docs/zh/guide/api.md
@@ -67,8 +67,8 @@ codeInspectorPlugin({
 ## injectTo <Badge type="tip" text="0.5.0+" vertical="middle" />
 
 - 可选项。默认值为 `auto`
-- 类型：`'auto' | 'all' | string`
-- 说明：用于注入 DOM 筛选和点击跳转 vscode 的相关的 `client code` 代码的文件(必须为绝对路径且以 `.js/.ts/.mjs/.mts/.jsx/.tsx` 为结尾的文件)。为 `auto` 时会在符合上述条件的第一个文件注入 `client code`；为 `all` 时会在每个符合上述条件的文件中都注入 `client code`；也可自己指定一个绝对路径文件作为注入 `client code` 的文件(通常用于 SSR 项目中指定一个 client 端的文件)。
+- 类型：`string`
+- 说明：用于注入 DOM 筛选和点击跳转 vscode 的相关的 `client code` 代码的文件(必须为绝对路径且以 `.js/.ts/.mjs/.mts/.jsx/.tsx` 为结尾的文件)。(通常用于 SSR 项目中指定一个 client 端的文件)。
 
 ## dev <Badge type="tip" text="0.5.0+" vertical="middle" />
