+++ packages/core/src/shared/constant.ts
@@ -2,16 +2,9 @@ export const LineName = 'data-insp-row'; // 注入的行名称
 export const ColumnName = 'data-insp-col'; // 注入的列名称
 export const PathName = 'data-insp-path'; // 注入的路径名称
 export const NodeName = 'data-insp-node'; // 注入的节点名称
-
-export type Editor =
-  | 'atom'
-  | 'code'
-  | 'code_insiders'
-  | 'idea'
-  | 'phpstorm'
-  | 'pycharm'
-  | 'webstorm'
-  | 'hbuilder';
-
 export const DefaultPort = 5678;
 export const JsFileExtList = ['.js', '.ts', '.mjs', '.mts', '.jsx', '.tsx'];
+export const ViteVirtualModule_EliminateVueWarning =
+  'virtual:code-inspector-plugin-eliminate-vue-warning-code';
+export const ViteVirtualModule_Client =
+  'virtual:code-inspector-plugin-client-code';
