+++ packages/core/src/shared/index.ts
@@ -1,2 +1,3 @@
 export * from './constant';
 export * from './utils';
+export * from './type';
