+++ packages/core/types/server/index.d.ts
@@ -1,4 +1,3 @@
-import { Editor } from '../shared/constant';
 export { transformCode } from './transform';
-export { getCodeWithWebComponent } from './inject-code';
-export declare function startServer(callback: (port: number) => any, editor?: Editor): void;
+export * from './use-client';
+export * from './server';
