+++ packages/core/types/server/launch-editor.d.ts
@@ -1,3 +1,3 @@
-import { Editor } from '../shared/constant';
+import { Editor } from '../shared';
 declare function launchEditor(fileName: string, lineNumber: unknown, colNumber: unknown, _editor?: Editor): void;
 export default launchEditor;
