+++ packages/core/types/server/use-client.d.ts
@@ -1,5 +1,5 @@
-import type { CodeOptions, RecordInfo } from '../index';
+import type { CodeOptions, RecordInfo } from '../shared';
 export declare const clientJsPath: string;
 export declare function getClientInjectCode(port: number, options?: CodeOptions): string;
-export declare const eliminateVueWarningCode: string;
+export declare function getEliminateVueWarningCode(): string;
 export declare function getCodeWithWebComponent(options: CodeOptions, file: string, code: string, record: RecordInfo): Promise<string>;
