+++ packages/core/types/shared/constant.d.ts
@@ -2,6 +2,7 @@ export declare const LineName = "data-insp-row";
 export declare const ColumnName = "data-insp-col";
 export declare const PathName = "data-insp-path";
 export declare const NodeName = "data-insp-node";
-export type Editor = 'atom' | 'code' | 'code_insiders' | 'idea' | 'phpstorm' | 'pycharm' | 'webstorm' | 'hbuilder';
 export declare const DefaultPort = 5678;
 export declare const JsFileExtList: string[];
+export declare const ViteVirtualModule_EliminateVueWarning = "virtual:code-inspector-plugin-eliminate-vue-warning-code";
+export declare const ViteVirtualModule_Client = "virtual:code-inspector-plugin-client-code";
