+++ packages/core/types/shared/index.d.ts
@@ -1,2 +1,3 @@
 export * from './constant';
 export * from './utils';
+export * from './type';
