+++ packages/core/types/shared/utils.d.ts
@@ -1,6 +1,6 @@
 export declare function fileURLToPath(fileURL: string): string;
 export declare function isJsTypeFile(file: string): boolean;
-export declare function isNextClientFile(code: string): boolean;
-export declare function isUseEffectFile(code: string): boolean;
+export declare function isNextJsEntry(code: string): boolean;
+export declare function isSsrEntry(code: string): boolean;
 export declare function getFilenameWithoutExt(filePath: string): string;
 export declare function normalizePath(filepath: string): string;
