+++ packages/vite-plugin/types/index.d.ts
@@ -6,6 +6,8 @@ export declare function ViteCodeInspectorPlugin(options?: Options): {
     apply(_: any, { command }: {
         command: any;
     }): boolean;
+    resolveId(id: any): Promise<string>;
+    load(id: any): string;
     transform(code: any, id: any): Promise<any>;
     enforce?: "pre";
     name: string;
